队列计数不能到达最长长度 修改外部中断触发-发送时不关闭外部中断 Signed-off-by: songbangyan <songbangyan@gmail.com>

+++ User/INC/myconfig.h
@@ -166,7 +166,7 @@
 typedef struct	
 	{   
 	u8 elem[MAXOFRXQ];   
-	u8 head,tail;   
+	u16 head,tail;   
 	}Quque;
 	
 #ifdef  root
